Project 1/Fetch Photo/requirements.txt
flask
redis
pymongo
python-dotenv
requests